+++ vector/src/vector_types/general/data_reorganization.rs
@@ -324,7 +324,7 @@ macro_rules! zero_interleave {
 }
 
 impl<S, T, N, D> InsertZerosOps<T> for DspVec<S, T, N, D>
-    where S: ToSliceMut<T> + Owner,
+    where S: ToSliceMut<T>,
           T: RealNumber,
           N: NumberSpace,
           D: Domain
@@ -332,7 +332,11 @@ impl<S, T, N, D> InsertZerosOps<T> for DspVec<S, T, N, D>
     fn zero_pad(&mut self, points: usize, option: PaddingOption) -> VoidResult {
         let len_before = self.len();
         let is_complex = self.is_complex();
-        let len = if is_complex { 2 * points } else { points };
+        let step = if is_complex { 2 } else { 1 };
+        let len = points * step;
+        if len <= len_before {
+            return Ok(());
+        }
 
         try!(self.resize(len));
         let data = self.data.to_slice_mut();
@@ -346,7 +350,7 @@ impl<S, T, N, D> InsertZerosOps<T> for DspVec<S, T, N, D>
                 Ok(())
             }
             PaddingOption::Surround => {
-                let diff = (len - len_before) / if is_complex { 2 } else { 1 };
+                let diff = (len - len_before) / step;
                 let mut right = (diff - 1) / 2;
                 let mut left = diff - right ;
                 if is_complex {
@@ -368,21 +372,23 @@ impl<S, T, N, D> InsertZerosOps<T> for DspVec<S, T, N, D>
                 Ok(())
             }
             PaddingOption::Center => {
-                let mut diff = (len - len_before) / if is_complex { 2 } else { 1 };
-                let mut left = diff / 2;
-                let mut right = diff - left;
+                let mut diff = (len - len_before) / step;
+                let points_before = len_before / step;
+                let mut right = (points_before - 1) / 2;
+                let mut left = points_before - right ;
                 if is_complex {
                     right *= 2;
                     left *= 2;
                     diff *= 2;
                 }
 
+                for i in 0..right {
+                    data[len - 1 - i] = data[len_before - 1 - i];
+                }
                 unsafe {
-                    let src = &data[left] as *const T;
-                    let dest = &mut data[len - right] as *mut T;
-                    ptr::copy(src, dest, right);
                     let dest = &mut data[left] as *mut T;
                     ptr::write_bytes(dest, 0, len - diff);
+                    
                 }
                 Ok(())
             }
@@ -483,8 +489,8 @@ impl<S, T, N, D> InsertZerosOpsBuffered<S, T> for DspVec<S, T, N, D>
                 }
                 PaddingOption::Center => {
                     let mut diff = (len - len_before) / if is_complex { 2 } else { 1 };
-                    let mut right = diff / 2;
-                    let mut left = diff - right;
+                    let mut right = (diff - 1) / 2;
+                    let mut left = diff - right ;
                     if is_complex {
                         right *= 2;
                         left *= 2;
@@ -492,9 +498,12 @@ impl<S, T, N, D> InsertZerosOpsBuffered<S, T> for DspVec<S, T, N, D>
                     }
 
                     unsafe {
-                        let src = &data[left] as *const T;
+                        let src = &data[len_before / 2 + 1] as *const T;
                         let dest = &mut target[len - right] as *mut T;
                         ptr::copy(src, dest, right);
+                        let src = &data[0] as *const T;
+                        let dest = &mut target[0] as *mut T;
+                        ptr::copy(src, dest, left);
                         let dest = &mut target[left] as *mut T;
                         ptr::write_bytes(dest, 0, len - diff);
                     }
@@ -651,11 +660,33 @@ mod tests {
     }
 
     #[test]
+    fn zero_pad_center_test() {
+        let mut v = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0].to_complex_time_vec();
+        v.zero_pad(10, PaddingOption::Center).unwrap();
+        let expected = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 
+            0.0, 0.0, 7.0, 8.0, 9.0, 10.0];
+        assert_eq!(&v[..], &expected);
+    }
+        
+
+    #[test]
+    fn zero_pad_b_center_test() {
+        let mut v = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0].to_complex_time_vec();
+        let mut buffer = SingleBuffer::new();
+        v.zero_pad_b(&mut buffer, 10, PaddingOption::Center);
+        let expected = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 
+            0.0, 0.0, 7.0, 8.0, 9.0, 10.0];
+        assert_eq!(&v[..], &expected);
+    }
+    
+    #[test]
     fn zero_pad_surround_odd_signal_test() {
-        let mut v = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0].to_complex_time_vec();
-        v.zero_pad(10, PaddingOption::Surround).unwrap();
-        let expected = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0,
-                        10.0, 11.0, 0.0, 0.0, 0.0];
+        let mut v = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0].to_real_time_vec();
+        v.zero_pad(20, PaddingOption::Surround).unwrap();
+        // The expected result is required so that the convolution theorem holds true 
+        // (mul in freq is the same as conv)
+        let expected = [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0,
+                        10.0, 11.0, 0.0, 0.0, 0.0, 0.0];
         assert_eq!(&v[..], &expected);
     }
 
@@ -681,11 +712,13 @@ mod tests {
 
     #[test]
     fn zero_pad_b_surround_odd_signal_test() {
-        let mut v = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0].to_complex_time_vec();
+        let mut v = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0].to_complex_time_vec();
         let mut buffer = SingleBuffer::new();
         v.zero_pad_b(&mut buffer, 10, PaddingOption::Surround);
+        // The expected result is required so that the convolution theorem holds true 
+        // (mul in freq is the same as conv)
         let expected = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0,
-                        10.0, 11.0, 0.0, 0.0, 0.0];
+                        10.0, 11.0, 12.0, 0.0, 0.0];
         assert_eq!(&v[..], &expected);
     }